requirements.txt
asgiref==3.8.1
black==24.10.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.5
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
